+++ remap/src/main/java/mapped/Class7396.java
@@ -65,6 +65,6 @@ public class Class7396
     }
     
     public static void method22725(final Color color) {
-        GL11.glColor4d((double)(color.getRed() / 255.0f), (double)(color.getGreen() / 255.0f), (double)(color.getBlue() / 255.0f), (double)(color.getAlpha() / 255.0f));
+        GL11.glColor4d(color.getRed() / 255.0f, color.getGreen() / 255.0f, color.getBlue() / 255.0f, color.getAlpha() / 255.0f);
     }
 }
